+++ previous-versions/common/src/codatcommon/sdkconfiguration.py
@@ -21,9 +21,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '0.35.0'
-    gen_version: str = '2.169.0'
-    user_agent: str = 'speakeasy-sdk/python 0.35.0 2.169.0 3.0.0 codat-common'
+    sdk_version: str = '0.35.1'
+    gen_version: str = '2.173.0'
+    user_agent: str = 'speakeasy-sdk/python 0.35.1 2.173.0 3.0.0 codat-common'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> Tuple[str, Dict[str, str]]: